feat: break index.d.ts into many interface files BREAKING CHANGES: all new structure and files
