+++ index.d.ts
@@ -1,387 +1,4 @@
-declare module 'pagarme' {
-  export namespace client {
-    /**
-     * Returns a version of client with authentication data binded to the resource requests.
-     *
-     * @example
-     * // API Key Authentication
-     * pagarme.client.connect({ api_key: 'ak_test_y7jk294ynbzf93' })
-     *
-     * // Encryption Key Authentication
-     * pagarme.client.connect({ encryption_key: 'ek_test_y7jk294ynbzf93' })
-     *
-     * // Login Authentication
-     * pagarme.client.connect({ email: '', password: '' })
-     *
-     * @param {Object} authentication
-     * @returns {Promise} A Promise that resolves to a client with authentication data binded
-     *
-     * References:
-     * - https://pagarme.github.io/pagarme-js/module-client.html#~connect
-     * - https://pagarme.github.io/pagarme-js/client_index.js.html#line88
-     */
-    function connect(authentication: {
-      api_key?: string;
-      encryption_key?: string;
-      email?: string;
-      password?: string;
-    }): Promise<typeof client>;
 
-    function search<T, Q extends SearchQuery>(
-      query: Q
-    ): Promise<
-      SearchOutput<
-        Q extends { type: 'customer' }
-          ? Customer
-          : Q extends { type: 'transaction' }
-          ? Transaction
-          : Q extends { type: 'subscription' }
-          ? Subscription
-          : Q extends { type: 'bank_account' }
-          ? BankAccount
-          : T
-      >
-    >;
-
-    function status(opts: any): any;
-
-    function versions(opts: any): any;
-
-    namespace acquirers {
-      function all(opts: any, pagination: any): any;
-
-      function create(opts: any, body: any): any;
-
-      function find(opts: any, body: any): any;
-
-      function findAll(a0: any, a1: any, ...args: any[]): any;
-
-      function update(opts: any, body: any): any;
-    }
-
-    namespace acquirersConfigurations {
-      function all(opts: any, pagination: any): any;
-
-      function create(opts: any, body: any): any;
-
-      function find(opts: any, body: any): any;
-
-      function findAll(a0: any, a1: any, ...args: any[]): any;
-
-      function update(opts: any, body: any): any;
-    }
-
-    namespace antifraudAnalyses {
-      function create(opts: any, body: any): any;
-
-      function find(opts: any, body: any): any;
-    }
-
-    namespace balance {
-      function find(opts: BalanceFindInput): Promise<BalanceObject>;
-
-      function primary(opts: any): any;
-    }
-
-    namespace balanceOperations {
-      function all(arg: FindAllBalanceOperations): Promise<BalanceOperation>;
-
-      function days(opts: any, body: any): any;
-
-      function find(opts: any, body: any): any;
-    }
-
-    namespace bankAccounts {
-      function all(opts: any, body: any): any;
-
-      function create(opts: CreateContaBancaria): Promise<ContaBancaria>;
-
-      function find(opts: any): Promise<ContaBancaria>;
-    }
-
-    namespace bulkAnticipations {
-      function all(opts: any, body: any): any;
-
-      function cancel(opts: any, body: any): any;
-
-      function confirm(opts: any, body: any): any;
-
-      function create(opts: any, body: any): any;
-
-      function days(opts: any, body: any): any;
-
-      function destroy(opts: any, body: any): any;
-
-      function find(opts: any, body: any): any;
-
-      function limits(opts: any, body: any): any;
-
-      function update(opts: any, body: any): any;
-    }
-
-    namespace cards {
-      function all(opts: any, pagination: any): any;
-
-      function create(opts: any, body: any): any;
-
-      function find(opts: any, body: any): any;
-    }
-
-    namespace chargebackOperations {
-      function find(a0: any, a1: any, ...args: any[]): any;
-    }
-
-    namespace chargebacks {
-      function find(opts: any, query: any): any;
-    }
-
-    namespace company {
-      function activate(opts: any): any;
-
-      function affiliationProgress(opts: any): any;
-
-      function create(opts: any, body: any): any;
-
-      function createTemporary(opts: any, body: any): any;
-
-      function current(opts: any): any;
-
-      function resetKeys(opts: any): any;
-
-      function update(opts: any, body: any): any;
-
-      function updateBranding(opts: any, body: any): any;
-
-      namespace emailTemplates {
-        function find(opts: any, body: any): any;
-
-        function update(opts: any, body: any): any;
-      }
-    }
-
-    namespace customers {
-      function all(opts: any, body: any): Promise<Customer[]>;
-
-      function create(body: CustomerInput): Promise<Customer>;
-
-      function find(opts: any, body: any): Promise<Customer>;
-    }
-
-    namespace events {
-      function find(opts: any, body: any): any;
-
-      function findCustom(opts: any, body: any): any;
-    }
-
-    namespace gatewayOperations {
-      function find(opts: any, body: any): any;
-
-      function refuseMessage(opts: any, body: any): any;
-    }
-
-    namespace invites {
-      function all(opts: any): any;
-
-      function create(opts: any, body: any): any;
-
-      function destroy(opts: any, body: any): any;
-
-      function find(opts: any, ...args: any[]): any;
-    }
-
-    namespace orders {
-      function all(opts: any, body: any): any;
-    }
-
-    namespace payables {
-      function all(args: FindRecebiveisArgs): Promise<Recebivel[]>;
-
-      function days(opts: any, body: any): any;
-
-      function find(opts: FindRecebivelArg): Promise<Recebivel[]>;
-    }
-
-    namespace paymentLinks {
-      function all(opts: any, body: any): any;
-
-      function cancel(opts: any, body: any): any;
-
-      function create(opts: any, body: any): any;
-
-      function find(opts: any, body: any): any;
-    }
-
-    namespace plans {
-      function all(opts: any, pagination: any): any;
-
-      function create(opts: any, body: any): any;
-
-      function find(opts: any, body: any): any;
-
-      function findAll(a0: any, a1: any, ...args: any[]): any;
-
-      function update(opts: any, body: any): any;
-    }
-
-    namespace postback {
-      function calculateSignature(
-        /** the keys used to sign the hash. */
-        key: string,
-        /** The string to be hashed. */
-        string: string
-      ): string;
-
-      function verifySignature(
-        /** the keys used to sign the hash. */
-        key: string,
-        /** The string to be hashed. */
-        string: string,
-        /** The expected result. */
-        expected: string
-      ): boolean;
-    }
-
-    namespace postbacks {
-      function find(opts: any, body: any): any;
-
-      function redeliver(opts: any, body: any): any;
-    }
-
-    namespace recipients {
-      function all(opts: any, body: any): any;
-
-      function create(opts: CreateRecebedor): Promise<Recebedor>;
-
-      function find(opts: any): any;
-
-      function update(opts: UpdateRecebedor): Promise<Recebedor>;
-    }
-
-    namespace refunds {
-      function find(opts: any, body: RefundsArgs): Promise<Refund[]>;
-      function cancel(opts: any, body: any): any;
-    }
-
-    namespace reprocessedTransactions {
-      function find(opts: any, query: any): any;
-    }
-
-    interface ICard {
-      card_holder_name: string;
-      card_expiration_date: string;
-      card_number: string;
-      card_cvv: string;
-    }
-
-    namespace security {
-      function encrypt(card: ICard): Promise<string>;
-
-      function sign(opts: any, string: any): any;
-
-      function verify(opts: any, string: any, expected: any): any;
-    }
-
-    namespace session {
-      function create(opts: any, email: any, password: any): any;
-
-      function destroy(opts: any, id: any): any;
-
-      function verify(opts: any, payload: any): any;
-    }
-
-    namespace splitRules {
-      function find(opts: any, body: any): any;
-    }
-
-    namespace subscriptions {
-      function all(opts: any, body: any): any;
-
-      function cancel(opts: any, body: any): any;
-
-      function create(opts: any, body: any): any;
-
-      function createTransaction(opts: any, body: any): any;
-
-      function find(opts: any, body: any): any;
-
-      function findAll(a0: any, a1: any, ...args: any[]): any;
-
-      function findTransactions(opts: any, body: any): any;
-
-      function settleCharge(opts: any, body: any): any;
-
-      function update(opts: any, body: any): any;
-    }
-
-    namespace transactions {
-      function all(opts: any, body: TransactionFindAll): Promise<Transaction[]>;
-
-      function calculateInstallmentsAmount(
-        opts: any,
-        body: CalculateInstallmentsAmountInput
-      ): Promise<CalculateInstallmentsAmount>;
-
-      function capture(opts: CaptureArgs): Promise<Transaction>;
-
-      function cardHashKey(opts: any): any;
-
-      function collectPayment(opts: any, body: any): any;
-
-      function create(opts: CreateTransactionInput): Promise<Transaction>;
-
-      function find<T extends TransactionFindArgs>(
-        opts: any,
-        body: T
-      ): Promise<T extends TransactionFindAll ? Transaction[] : Transaction>;
-
-      function refund(
-        opts: any,
-        body: TransactionRefundArgs
-      ): Promise<Transaction>;
-
-      function reprocess(opts: any, body: any): any;
-
-      function update(opts: any, body: any): any;
-    }
-
-    namespace transfers {
-      function all(args: FindAllTransfersArgs): Promise<TransferenciaObject[]>;
-
-      function cancel(opts: any, body: any): any;
-
-      function create(opts: TransferenciaInput): Promise<TransferenciaObject>;
-
-      function days(opts: any): any;
-
-      function find(opts: any, body: any): any;
-
-      function limits(opts: any, params: any): any;
-    }
-
-    namespace user {
-      function all(opts: any, body: any): any;
-
-      function create(opts: any, body: any): any;
-
-      function current(opts: any): any;
-
-      function destroy(opts: any, body: any): any;
-
-      function find(opts: any, body: any): any;
-
-      function redefinePassword(opts: any, body: any): any;
-
-      function resetPassword(opts: any, body: any): any;
-
-      function update(opts: any, body: any): any;
-
-      function updatePassword(opts: any, body: any): any;
-    }
-
-    namespace zipcodes {
-      function find(a0: any, a1: any, ...args: any[]): any;
-    }
-  }
 
   /**
    * This method validates the properties supplied in the object.
@@ -421,19 +38,7 @@ declare module 'pagarme' {
    * - https://pagarme.github.io/pagarme-js/module-validations.html
    * - https://pagarme.github.io/pagarme-js/validations_index.js.html#line69
    */
-  export function validate(body: {
-    cnpj?: string | string[] | number | number[];
-    cpf?: string | string[] | number | number[];
-    ddd?: string | string[] | number | number[];
-    zipcode?: string | string[] | number | number[];
-    phone?: string | string[] | number | number[];
-    card?: {
-      card_holder_name: string;
-      card_number: string | number;
-      card_cvv: string | number;
-      card_expiration_date: string | number;
-    };
-  }): {
+  export function validate(body: ): {
     cnpj?: boolean;
     cpf?: boolean;
     ddd?: boolean;
@@ -460,349 +65,6 @@ declare module 'pagarme' {
     /** Filtro pela data de atualização do estorno */
     date_updated?: string;
   }
-  export interface Refund {
-    object: 'refund';
-    /** Número identificador do estorno. */
-    id: string;
-    /** Valor, em centavos, do estorno. Exemplo: R$100,00 = 10000 */
-    amount: number;
-    /** Tipo de pagamento da transação estornada. */
-    type: 'credit_card' | 'debit_card' | 'boleto';
-    /** Representa o estado do estorno. */
-    status: 'refunded' | 'pending_refund';
-    /** Número identificador do recebedor responsável pela taxa de processamento. */
-    charge_fee_recipient_id: string;
-    /** Número identificador da conta bancária de destino do estorno. */
-    bank_account_id: number;
-    /** Número identificador da transação estornada. */
-    transaction_id: number;
-    /** Data de criação do estorno no formato ISODate */
-    date_created: string;
-    /** Objeto com dados adicionais informados na criação do estorno. */
-    metadata: JSON;
-  }
-
-  export interface Address {
-    /** País. Duas letras minúsculas. Deve seguir o padrão `ISO 3166-1 alpha-2` */
-    country: string;
-    /** Estado */
-    state: string;
-    /** Cidade */
-    city: string;
-    /** Rua */
-    street: string;
-    /** Número */
-    street_number: string;
-    /** Cidade */
-    zipcode: string;
-    /** Bairro */
-    neighborhood?: string;
-    /** Complemento. **Não pode ser uma string vazia** nem null */
-    complementary: string;
-  }
-
-  export type CustomerType = 'individual' | 'corporation' | 'other';
-  export type DocumentType = 'cpf' | 'cnpj' | 'passaporte' | 'other';
-
-  export interface Document {
-    /** Tipo de documento. Para compradores brasileiros, deve ser fornecido ao menos um CPF (no caso de pessoa física, i.e. `individual`) ou CNPJ (no caso de pessoa jurídica, i.e. `corporation`). Para compradores internacionais, o documento pode ser um passaporte (type `passport`) ou um campo personalizado (type `other`). */
-    type: DocumentType;
-    /** Número do documento */
-    number: string;
-  }
-
-  export interface CustomerInput {
-    external_id: string;
-    /** Nome ou razão social do comprador */
-    name: string;
-    /** Tipo de documento. Deve ser `individual` para pessoa física ou `corporation` para pessoa jurídica */
-    type: CustomerType;
-    /** País */
-    country: Country;
-    /** E-mail do comprador */
-    email: string;
-    /** Documento. Contém campos type para tipo de documento e number para número do documento. */
-    documents: Document[];
-    /** Números de telefone. Requer ao menos um valor. Deve seguir o padrão *E.164* */
-    phone_numbers: string[];
-    /** Data de nascimento */
-    birthday?: string;
-  }
-
-  export interface Customer extends CustomerInput {
-    /** Identificador do cliente na loja */
-    id: string;
-    /** Lista dos telefones relacionados ao cliente */
-    phones: string[];
-    /** Lista de endereços relacionados ao cliente */
-    addresses: Address[];
-  }
-
-  export interface ShippingInput {
-    /** Nome da entidade de cobrança */
-    name: string;
-    /** Taxa de envio cobrada do comprador. Por exemplo, se a taxa de envio é de dez reais e três centavos (R$10,03), o valor deve ser fornecido como ‘1003’ */
-    fee: number;
-    /** Data de entrega. Estimativa fornecida no formato AAAA-MM-DD */
-    delivery_date?: string;
-    /** Entrega expressa. Se for entrega expressa, deve conter ‘true’ (sim). Caso contrário, deve conter ‘false’ (não) */
-    expedited?: boolean;
-    /** Obrigatório. Dados do endereço de envio. Objeto descrito aqui. */
-    address: Address;
-  }
-
-  export interface ItemInput {
-    /** SKU (unidade de manutenção de estoque) ou número de identificação na loja */
-    id: string;
-    /** Nome do item vendido. */
-    title: string;
-    /** Preço por unidade. Por exemplo, se o preço de cada item é vinte reais e seis centavos (R$20,06), o valor deve ser fornecido como ‘2006’ */
-    unit_price: number;
-    /** Número de unidades vendidas do produto */
-    quantity: number;
-    /** Caracteriza o produto como bem físico ou não. Por bem físico, entende-se produtos que devem ser enviados fisicamente ao comprador, como calçados, eletrônicos e brinquedos. Se for um bem físico deve conter true (sim). Caso contrário, deve conter false (não) */
-    tangible: boolean;
-    /** Categoria */
-    category?: string;
-    /** Local */
-    venue?: string;
-    /** Data Estimativa fornecida no formato AAAA-MM-DD */
-    date?: string;
-  }
-
-  export interface CreateTransactionPixInput {
-    payment_method: 'pix';
-    /** Deve ser informada a data de expiração do Pix. Aceitaremos os
-    formatos ISO 8601 (YYYY-MM-DD) para Pix com expiração até o fim do dia e RFC3339
-    (YYYY-MM-DDThh:mm:ss) quando o Pix deva ser encerrado em algum horário específico
-    do dia */
-    pix_expiration_date: string;
-    /** campos que serão exibidos como soft descriptors ao pagador */
-    pix_additional_fields: {
-      name: string;
-      value: string;
-    }[];
-  }
-
-  export interface CreateTransactionBoletoInput {
-    payment_method: 'boleto';
-    boleto_fine?: {
-      /** Dias após a expiração do boleto quando a multa deve ser cobrada. */
-      days?: string;
-      /** Valor em centavos da multa. **Valor máximo de 2% do valor do documento.** */
-      amount?: string;
-    };
-    boleto_interest?: {
-      /** Dias após a expiração do boleto quando o juros deve ser cobrado. */
-      days?: string;
-      /** Valor em porcentagem da taxa de juros que será cobrado por dia. **Valor máximo de 1% ao mês.** */
-      amount?: string;
-    };
-    /** Prazo limite para pagamento do boleto. Deve ser passado no formato yyyy-MM-dd.Default: data atual + 7 dias */
-    boleto_expiration_date?: string;
-    /** Campo instruções do boleto. Máximo de 255 caracteres */
-    boleto_instructions?: string;
-  }
-
-  export interface CreateTransactionCreditCartInputBase {
-    payment_method: 'credit_card';
-  }
-
-  export interface CreateTransactionCreditCartInputWithHash {
-    /** Informações do cartão do cliente criptografadas em sua aplicação. **OBS: apenas para transações de Cartão de crédito você deve passar o `card_hash` ou `card_id`. Caso inclua os dados do cartão diretamente pelo código, esse campo torna-se dispensável.** */
-    card_hash: string;
-  }
-
-  export interface CreateTransactionCreditCartInputWitId {
-    /** Ao realizar uma transação, retornamos o card_id do cartão, para que nas próximas transações ele possa ser utilizado como forma de identificar os dados de pagamento. Exemplo de utilização: One-click buy. OBS: apenas para transações de Cartão de crédito você deve passar o card_hash ou card_id. Caso inclua os dados do cartão diretamente pelo código, esse campo torna-se dispensável. */
-    card_id: string;
-  }
-
-  export interface CreateTransactionCreditCartInputWitData {
-    /** Nome do portador do cartão.  */
-    card_holder_name: string;
-    /** Data de validade do cartão no formato MMAA. */
-    card_expiration_date: string;
-    /** Número do cartão. */
-    card_number: string;
-    /** Código verificador do cartão. */
-    card_cvv: string;
-  }
-
-  export type CreateTransactionCreditCartInput = CreateTransactionCreditCartInputBase &
-    (
-      | CreateTransactionCreditCartInputWitData
-      | CreateTransactionCreditCartInputWithHash
-      | CreateTransactionCreditCartInputWitId
-    );
-
-  interface CreateTransactionInputBase {
-    /** Valor a ser cobrado. Deve ser passado em centavos. Ex: R$ 10.00 = 1000. Deve ser no mínimo 1 real (100) */
-    amount: number;
-    /** Endpoint do seu sistema que receberá informações a cada atualização da transação. **Caso você defina este parâmetro, o processamento da transação se torna assíncrono.** */
-    postback_url?: string;
-    /** Utilize false caso queira manter o processamento síncrono de uma transação. Ou seja, a resposta da transação é recebida na hora. */
-    async?: boolean;
-    /** Número de parcelas da transação, sendo mínimo: 1 e Máximo: 12. OBS: Se o pagamento for boleto, o padrão é 1 */
-    installments?: string;
-    /** Descrição que aparecerá na fatura depois do nome de sua empresa. Máximo de 13 caracteres, sendo alfanuméricos e espaços. */
-    soft_descriptor?: string;
-    /** Após a autorização de uma transação, você pode escolher se irá capturar ou adiar a captura do valor. Caso opte por postergar a captura, atribua o valor false. */
-    capture?: boolean;
-    // TODO: Finalizar tipagem
-    /** Regras de divisão da transação */
-    split_rules?: Array<any>;
-    customer?: CustomerInput | { id: string };
-    /** Obrigatório com o antifraude habilitado. Define os dados de cobrança, como nome e endereço */
-    billing?: BillingInput;
-    /** Deve ser preenchido no caso da venda de bem físico (ver objeto items) */
-    shipping?: ShippingInput;
-    /** Define os dados dos itens vendidos, como nome, preço unitário e quantidade */
-    items?: ItemInput[];
-    /** Você pode passar dados adicionais na criação da transação para facilitar uma futura análise de dados tanto em nossa dashboard, quanto por seus sistemas. Ex: metadata[ idProduto ]=13933139 */
-    metadata?: string;
-    /** Valor único que identifica a transação para permitir uma nova tentativa de requisição com a segurança de que a mesma operação não será executada duas vezes acidentalmente.*/
-    reference_key?: string;
-    /** Valor único que identifica a sessão do usuário acessando o site. Máximo de 100 caracteres */
-    session?: string;
-    /** Data e hora do dispositivo que está efetuando a transação. Deve ser enviado no seguinte formato: yyyy-MM-dd'T'HH:mm:ss'Z. Por exemplo: 2017-10-31T14:53:00.000Z. OBS.: este campo é necessário para transações de mundo físico (com método de captura EMV e Magstripe) */
-    local_time?: string;
-  }
-
-  export type CreateTransactionInput = CreateTransactionInputBase &
-    (
-      | CreateTransactionCreditCartInput
-      | CreateTransactionBoletoInput
-      | CreateTransactionPixInput
-    );
-
-  type RefuseStatus =
-    | 'acquirer'
-    | 'antifraud'
-    | 'internal_error'
-    | 'no_acquirer'
-    | 'acquirer_timeout';
-
-  type TransactionStatus =
-    | 'processing'
-    | 'authorized'
-    | 'paid'
-    | 'refunded'
-    | 'waiting_payment'
-    | 'pending_refund'
-    | 'refused'
-    | 'chargedback'
-    | 'analyzing'
-    | 'pending_review';
-
-  interface Transaction {
-    /** Nome do tipo do objeto criado/modificado. */
-    object: 'transaction';
-    /** Representa o estado da transação. A cada atualização no processamento da transação, esta propriedade é alterada e, caso você esteja usando uma postback_url, os seus servidores são notificados desses updates. */
-    status: TransactionStatus;
-    /** Motivo pelo qual a transação foi recusada. */
-    refuse_reason?: RefuseStatus;
-    /** Agente responsável pela validação ou anulação da transação. */
-    status_reason: RefuseStatus;
-    /** Adquirente responsável pelo processamento da transação. */
-    acquirer_name: 'development' | 'pagarme' | 'stone' | 'cielo' | 'rede';
-    /** ID da adquirente responsável pelo processamento da transação. */
-    acquirer_id: string;
-    /** Mensagem de resposta da adquirente referente ao status da transação. */
-    acquirer_response_code: string;
-    /** Código de autorização retornado pela bandeira. */
-    authorization_code: string;
-    /** Texto que irá aparecer na fatura do cliente depois do nome da loja. */
-    soft_descriptor: string;
-    /** Código que identifica a transação na adquirente. */
-    tid: string;
-    /** Código que identifica a transação na adquirente. */
-    nsu: string;
-    /** Data de criação da transação no formato ISODate */
-    date_created: string;
-    /** Data de atualização  da transação no formato ISODate */
-    date_updated: string;
-    /** Valor, em centavos, da transação. Ex: 100,00 = 10000 */
-    amount: number;
-    /** Valor em centavos autorizado na transação, sempre menor ou igual a `amount`. */
-    authorized_amount: number;
-    /** Valor em centavos capturado na transação, sempre menor ou igual a `authorized_amount`. */
-    paid_amount: number;
-    /** Valor em centavos estornado até o momento na transação, sempre menor ou igual a `paidamount`. */
-    refunded_amount: number;
-    /** Número de parcelas a serem cobradas. */
-    installments: number;
-    /** Número identificador da transação */
-    id: number;
-    /** Custo da transação para o lojista, envolvendo processamento e antifraude. */
-    cost: number;
-    /** Nome do portador do cartão. */
-    card_holder_name: string;
-    /** Últimos 4 dígitos do cartão. */
-    card_last_digits: string;
-    /** ÚPrimeiros 5 dígitos do cartão */
-    card_first_digits: string;
-    /** Bandeira do cartão. */
-    card_brand: string;
-    /** Usado em transações EMV, define se a validação do cartão aconteceu online(com banco emissor), ou offline( através do chip). */
-    card_pin_mode: string;
-    /** URL (endpoint) de seu sistema que recebe notificações a cada mudança no status da transação. */
-    postback_url: string;
-    /** Método de pagamento */
-    payment_method: 'credit_card' | 'boleto' | 'pix';
-    pix_qrcode: string;
-    pix_expiration_date: string;
-    pix_additional_fields: string;
-    /** Define qual foi a forma de captura dos dados de pagamento. */
-    capture_method: 'magstripe' | 'emv' | 'ecommerce';
-    /** Define qual foi a nota de antifraude atribuída a transação. Lembrando que por padrão, transações com score >= 95 são recusadas. */
-    antifraud_score: string;
-    /** URL do boleto para impressão */
-    boleto_url: string;
-    /** Código de barras do boleto gerado na transação */
-    boleto_barcode: string;
-    /** Data de expiração do boleto (em ISODate) */
-    boleto_expiration_date: string;
-    /** Mostra se a transação foi criada utilizando a API Key ou Encryption Key */
-    referer: string;
-    /** IP de origem que criou a transação, podendo ser diretamente de seu cliente, caso a requisição venha diretamente do client-side, ou de seus servidores, caso tudo esteja centralizando em sua aplicação no server-side. */
-    ip: string;
-    /** Caso essa transação tenha sido originada na cobrança de uma assinatura, o id desta será o valor dessa propriedade. */
-    subscription_id: string;
-    customer: CustomerInput;
-    billing: BillingInput;
-    shipping: ShippingInput;
-    items: ItemInput[];
-    address: Address;
-    documents: Document[];
-    /** Objeto com dados adicionais informados na criação da transação. */
-    metadata: any;
-    /** Objeto com as regras de split definidas para essa transação. */
-    split_rules: any;
-    /** Objeto com dados usados na integração com antifraude. */
-    antifraud_metadata: any;
-    /** Valor único que identifica a sessão do usuário acessando o site */
-    session: string;
-    /** Valor único que identifica a transação para permitir uma nova tentativa de requisição com a segurança de que a mesma operação não será executada duas vezes acidentalmente. */
-    reference_key: string;
-  }
-  // TODO: Atualizar tipagem transaction | subscription
-  export interface Postback {
-    /** ID da transação. */
-    id: number;
-    /** A qual evento o postback se refere.  */
-    event: 'transaction_status_changed' | 'subscription_status_changed';
-    /** Status anterior da transação. */
-    old_status: TransactionStatus;
-    /** Status ideal para objetos deste tipo, em um fluxo normal, onde autorização e captura são feitos com sucesso, por exemplo. */
-    desired_status: TransactionStatus;
-    /** Status para o qual efetivamente mudou. */
-    current_status: TransactionStatus;
-    /** Qual o tipo do objeto referido.  */
-    object: 'transaction' | 'subscription';
-    /** Possui todas as informações do objeto.  */
-    transaction: Transaction;
-  }
 
   interface TransactionRefundBoletoDataArgs {
     /** Objeto bank_account que contém os dados da conta bancária para onde o estorno será feito. */
@@ -819,10 +81,10 @@ declare module 'pagarme' {
       conta_dv: string;
       /** Tipo da conta bancária. */
       type:
-      | 'conta_corrente'
-      | 'conta_poupanca'
-      | 'conta_corrente_conjunta'
-      | 'conta_poupanca_conjunta';
+        | 'conta_corrente'
+        | 'conta_poupanca'
+        | 'conta_corrente_conjunta'
+        | 'conta_poupanca_conjunta';
       /** CPF ou CNPJ do favorecido */
       document_number: string;
       /** Nome/razão social do favorecido, Até 30 caracteres */
@@ -834,7 +96,7 @@ declare module 'pagarme' {
     bank_account_id: string;
   }
   // eslint-disable-next-line @typescript-eslint/no-empty-interface
-  interface TransactionRefundCreditCardArgs { }
+  interface TransactionRefundCreditCardArgs {}
 
   export interface TransactionRefundDefaultArgs {
     /** The transaction ID. */
@@ -902,7 +164,7 @@ declare module 'pagarme' {
   interface CalculateInstallmentsAmount {
     installments: {
       [key: string]: Installment;
-    }
+    };
   }
 
   interface Installment {
@@ -919,11 +181,6 @@ declare module 'pagarme' {
 
   export interface Authentication {
     api_key: string;
-  }
-
-  export interface BillingInput {
-    name: string;
-    address: Address;
   }
 
   // CONTA BANCARIA
@@ -1048,13 +305,13 @@ declare module 'pagarme' {
     payment_date: string;
     original_payment_date: string;
     type:
-    | 'credit'
-    | 'refund'
-    | 'refund_reversal'
-    | 'chargeback'
-    | 'chargeback_refund'
-    | 'block'
-    | 'unblock';
+      | 'credit'
+      | 'refund'
+      | 'refund_reversal'
+      | 'chargeback'
+      | 'chargeback_refund'
+      | 'block'
+      | 'unblock';
     payment_method: 'credit_card' | 'debit_card' | 'boleto';
     accrual_date: string;
     date_created: string;
@@ -1090,11 +347,11 @@ declare module 'pagarme' {
     Amount: number | string;
     Type: 'ted' | 'doc' | 'credito_em_conta';
     Status:
-    | 'pending_transfer'
-    | 'transferred'
-    | 'failed'
-    | 'processing'
-    | 'canceled';
+      | 'pending_transfer'
+      | 'transferred'
+      | 'failed'
+      | 'processing'
+      | 'canceled';
     Fee: number | string;
     Funding_date: string;
     Funding_estimated_date: string;
@@ -1171,251 +428,5 @@ declare module 'pagarme' {
     type: 'transaction' | 'subscription' | 'bank_account' | 'customer';
     query: Record<string, unknown> | string;
     search_type?: string;
-  }
-
-  enum Country {
-    Af = 'af',
-    Al = 'al',
-    Dz = 'dz',
-    As = 'as',
-    Ad = 'ad',
-    Ao = 'ao',
-    Ai = 'ai',
-    Aq = 'aq',
-    Ag = 'ag',
-    Ar = 'ar',
-    Am = 'am',
-    Aw = 'aw',
-    Au = 'au',
-    At = 'at',
-    Az = 'az',
-    Bs = 'bs',
-    Bh = 'bh',
-    Bd = 'bd',
-    Bb = 'bb',
-    By = 'by',
-    Be = 'be',
-    Bz = 'bz',
-    Bj = 'bj',
-    Bm = 'bm',
-    Bt = 'bt',
-    Bo = 'bo',
-    Ba = 'ba',
-    Bw = 'bw',
-    Bv = 'bv',
-    Br = 'br',
-    Io = 'io',
-    Bn = 'bn',
-    Bg = 'bg',
-    Bf = 'bf',
-    Bi = 'bi',
-    Kh = 'kh',
-    Cm = 'cm',
-    Ca = 'ca',
-    Cv = 'cv',
-    Ky = 'ky',
-    Cf = 'cf',
-    Td = 'td',
-    Cl = 'cl',
-    Cn = 'cn',
-    Cx = 'cx',
-    Cc = 'cc',
-    Co = 'co',
-    Km = 'km',
-    Cg = 'cg',
-    Cd = 'cd',
-    Ck = 'ck',
-    Cr = 'cr',
-    Ci = 'ci',
-    Hr = 'hr',
-    Cu = 'cu',
-    Cy = 'cy',
-    Cz = 'cz',
-    Dk = 'dk',
-    Dj = 'dj',
-    Dm = 'dm',
-    Do = 'do',
-    Ec = 'ec',
-    Eg = 'eg',
-    Sv = 'sv',
-    Gq = 'gq',
-    Er = 'er',
-    Ee = 'ee',
-    Et = 'et',
-    Fk = 'fk',
-    Fo = 'fo',
-    Fj = 'fj',
-    Fi = 'fi',
-    Fr = 'fr',
-    Gf = 'gf',
-    Pf = 'pf',
-    Tf = 'tf',
-    Ga = 'ga',
-    Gm = 'gm',
-    Ge = 'ge',
-    De = 'de',
-    Gh = 'gh',
-    Gi = 'gi',
-    Gr = 'gr',
-    Gl = 'gl',
-    Gd = 'gd',
-    Gp = 'gp',
-    Gu = 'gu',
-    Gt = 'gt',
-    Gg = 'gg',
-    Gn = 'gn',
-    Gw = 'gw',
-    Gy = 'gy',
-    Ht = 'ht',
-    Hm = 'hm',
-    Va = 'va',
-    Hn = 'hn',
-    Hk = 'hk',
-    Hu = 'hu',
-    Is = 'is',
-    In = 'in',
-    Id = 'id',
-    Ir = 'ir',
-    Iq = 'iq',
-    Ie = 'ie',
-    Im = 'im',
-    Il = 'il',
-    It = 'it',
-    Jm = 'jm',
-    Jp = 'jp',
-    Je = 'je',
-    Jo = 'jo',
-    Kz = 'kz',
-    Ke = 'ke',
-    Ki = 'ki',
-    Kp = 'kp',
-    Kr = 'kr',
-    Kw = 'kw',
-    Kg = 'kg',
-    La = 'la',
-    Lv = 'lv',
-    Lb = 'lb',
-    Ls = 'ls',
-    Lr = 'lr',
-    Ly = 'ly',
-    Li = 'li',
-    Lt = 'lt',
-    Lu = 'lu',
-    Mo = 'mo',
-    Mk = 'mk',
-    Mg = 'mg',
-    Mw = 'mw',
-    My = 'my',
-    Mv = 'mv',
-    Ml = 'ml',
-    Mt = 'mt',
-    Mh = 'mh',
-    Mq = 'mq',
-    Mr = 'mr',
-    Mu = 'mu',
-    Yt = 'yt',
-    Mx = 'mx',
-    Fm = 'fm',
-    Md = 'md',
-    Mc = 'mc',
-    Mn = 'mn',
-    Me = 'me',
-    Ms = 'ms',
-    Ma = 'ma',
-    Mz = 'mz',
-    Mm = 'mm',
-    Na = 'na',
-    Nr = 'nr',
-    Np = 'np',
-    Nl = 'nl',
-    An = 'an',
-    Nc = 'nc',
-    Nz = 'nz',
-    Ni = 'ni',
-    Ne = 'ne',
-    Ng = 'ng',
-    Nu = 'nu',
-    Nf = 'nf',
-    Mp = 'mp',
-    No = 'no',
-    Om = 'om',
-    Pk = 'pk',
-    Pw = 'pw',
-    Ps = 'ps',
-    Pa = 'pa',
-    Pg = 'pg',
-    Py = 'py',
-    Pe = 'pe',
-    Ph = 'ph',
-    Pn = 'pn',
-    Pl = 'pl',
-    Pt = 'pt',
-    Pr = 'pr',
-    Qa = 'qa',
-    Re = 're',
-    Ro = 'ro',
-    Ru = 'ru',
-    Rw = 'rw',
-    Sh = 'sh',
-    Kn = 'kn',
-    Lc = 'lc',
-    Pm = 'pm',
-    Vc = 'vc',
-    Ws = 'ws',
-    Sm = 'sm',
-    St = 'st',
-    Sa = 'sa',
-    Sn = 'sn',
-    Rs = 'rs',
-    Sc = 'sc',
-    Sl = 'sl',
-    Sg = 'sg',
-    Sk = 'sk',
-    Si = 'si',
-    Sb = 'sb',
-    So = 'so',
-    Za = 'za',
-    Gs = 'gs',
-    Es = 'es',
-    Lk = 'lk',
-    Sd = 'sd',
-    Sr = 'sr',
-    Sj = 'sj',
-    Sz = 'sz',
-    Se = 'se',
-    Ch = 'ch',
-    Sy = 'sy',
-    Tw = 'tw',
-    Tj = 'tj',
-    Tz = 'tz',
-    Th = 'th',
-    Tl = 'tl',
-    Tg = 'tg',
-    Tk = 'tk',
-    To = 'to',
-    Tt = 'tt',
-    Tn = 'tn',
-    Tr = 'tr',
-    Tm = 'tm',
-    Tc = 'tc',
-    Tv = 'tv',
-    Ug = 'ug',
-    Ua = 'ua',
-    Ae = 'ae',
-    Gb = 'gb',
-    Us = 'us',
-    Um = 'um',
-    Uy = 'uy',
-    Uz = 'uz',
-    Vu = 'vu',
-    Ve = 've',
-    Vn = 'vn',
-    Vg = 'vg',
-    Vi = 'vi',
-    Wf = 'wf',
-    Eh = 'eh',
-    Ye = 'ye',
-    Zm = 'zm',
-    Zw = 'zw'
   }
 }
