+++ src/common/CustomerType.d.ts
@@ -1,0 +1,1 @@
+export type CustomerType = 'individual' | 'corporation' | 'other';
