+++ src/common/PaymentMethod.d.ts
@@ -1,0 +1,1 @@
+export type PaymentMethod = 'credit_card' | 'boleto' | 'pix';
