+++ src/common/RefuseStatus.d.ts
@@ -1,0 +1,6 @@
+export type RefuseStatus =
+  | 'acquirer'
+  | 'antifraud'
+  | 'internal_error'
+  | 'no_acquirer'
+  | 'acquirer_timeout';
