+++ src/common/TransactionStatus.d.ts
@@ -1,0 +1,11 @@
+export type TransactionStatus =
+  | 'processing'
+  | 'authorized'
+  | 'paid'
+  | 'refunded'
+  | 'waiting_payment'
+  | 'pending_refund'
+  | 'refused'
+  | 'chargedback'
+  | 'analyzing'
+  | 'pending_review';
