+++ src/inputs/InputBilling.d.ts
@@ -1,0 +1,4 @@
+export interface InputBilling {
+  name: string;
+  address: Address;
+}
