+++ src/inputs/Transaction/CreditCard/index.d.ts
@@ -1,0 +1,3 @@
+export * from './CreditCard';
+export * from './CreditCardByHash';
+export * from './CreditCardById';
