+++ src/inputs/Transaction/index.d.ts
@@ -1,0 +1,3 @@
+export * from './InputTransactionPix';
+export * from './InputTransactionBoleto';
+export * from './InputTransactionCreditCard';
