+++ src/inputs/index.ts
@@ -1,0 +1,5 @@
+export * from './InputItem.d';
+export * from './InputShipping.d';
+export * from './InputBilling.d';
+export * from './InputCustomer.d';
+export * from './InputTransaction.d';
